% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{sqlconvertcall}
\alias{sqlconvertcall}
\title{Translate a SQL file}
\usage{
sqlconvertcall(sql = "", targetDialect,
  oracleTempSchema = NULL)
}
\arguments{
\item{sql}{The source SQL file}


\item{targetDialect}{The target dialect. Currently 'oracle', 'postgresql', and 'redshift' are
supported}

\item{oracleTempSchema}{A schema that can be used to create temp tables in when using Oracle.}
}
\description{
This function takes SQL and translates it to a different dialect.
}
\details{
This function takes SQL and translates it to a different dialect.
}
\examples{
\dontrun{
sqlconvertcall("myRenderedStatement.sql",
                 "myTranslatedStatement.sql",
                 targetDialect = "postgresql")
}
}
